Add nedb database and client sessions
server.ts now uses a nedb database to store users.
backend has a db directory with class User, which represents the site's
users and a db admin module that performs operations on the database.
React App now has a simple login screen in which the user can request to
log into the server.
addUser.ts is a simple command line utility that can be invoked via yarn
scripts that adds a new user to the database.

tsconfig now uses commonjs modules