Implement bookmarks
NavBar now shows a bookmark button in the ChapterPage route.
When clicked sends a request to server to store the scrollPosition on
database so that next time that chapter is fetched, the app scrolls to
the last position.

User class now has a bookmark object where each bookmark fraction value
is stored with different keys following a defined pattern.

When authentication errors arise, server returns status code 401 and the
client has a new function unauthorizedResponseHandler that reloads the
page so that the client receives a new bundle without user data.

Fixed bug where server crashed when trying to serve app to an user
without cookies.