New page for sending feedback
Settings menu has a new button: send feedback. This links to the new
route: /feedback where the user can send a feedback message. This
message is sent to the server and then forwarded via telegram.

The LoginError component from Login is now in the comp dir as FormError
so that it can be shared with the feedback page.

RectangularIconButton now centers text if no icon is provided.