Host videos on the server
Instead of relying on youtube, it is now possible to host videos on the
server. Videos should be placed on backend/videos. On the lyricsIndex,
song can have an optional field 'local', if that field is present then
the video is played from the server rather than Youtube.

There are 2 new components for video playback: YoutubePlayer and
LocalPlayer. LyricsPage.js decides which to use based on the contents of
the song object as described in the previous paragraph.