Use React Router links in WhatsNewBox
Markdown links in WhatsNewBox now use React Router to avoid triggering a
page refresh.
PreloadedStateQueries now validades that bookIndex, lyricsIndex and
changelog exist, because if server doesn't render them, the initial page
load would crash.
RouteTitleProps now validates that the user is logged in becase if it
isn't and it tries to fetch data for a book or song, it will crash
because PreloadedState does not contain them